Height measurement device
Mount ultrasonic sensor SR04 to the ceiling and wire it to the arduino.

The arduino should have a 16x2 lcd screen and 3 additional buttons wired for reset, measure, and convert functions.

Once the board is installed on the wall and sensor on the ceiling, press button 1, reset, to take a measurement to the floor with no obstacles. Stand under the sensor and press button 2 to measure the distance to the top of your height on the screen. Button 3 is used to convert between cm, in, and feet/in